servicio-2/requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.86.0
h11==0.14.0
idna==3.4
influxdb-client==1.34.0
msgpack==1.0.4
paho-mqtt==1.6.1
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.6
reactivex==4.0.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
